[Refactor][JUnit] Updated JUnit utils due to split in command classes
