+++ src/test/java/seedu/address/testutil/TaskUtil.java
@@ -7,7 +7,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PRIORITY;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TASK;
 
-import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.AddTaskCommand;
 import seedu.address.model.task.ReadOnlyTask;
 
 /**
@@ -19,7 +19,7 @@ public class TaskUtil {
      * Returns an add command string for adding the {@code task}.
      */
     public static String getAddCommand(ReadOnlyTask task) {
-        return AddCommand.COMMAND_WORD + " " + PREFIX_TASK + " " + getTaskDetails(task);
+        return AddTaskCommand.COMMAND_WORD + " " + PREFIX_TASK + " " + getTaskDetails(task);
     }
 
     /**
